fix: extract item field helpers to resolve CI audit duplication and complexity violations
Extract AnyItem type and five getItem* helpers into ai_config.helpers.ts so
ItemCard and ItemListRow share the derivation logic instead of duplicating it.
Add fallow-ignore-next-line complexity suppressions for unavoidable
discriminator functions in dialogs.
